qt-6: modify the geoclue dependency to geoclue2
What QtPositioning uses is really GeoClue2 instead of legacy GeoClue.

In addition, this dependency may be reduced to recommendation because
the QtPositioning GeoClue2 plugin communicates with GeoClue2 via D-Bus
instead of linking to it.

Jumbo build is disable for QtWebEngine on LoongArch64 to prevent build
failure, and a patch is added for fixing non-jumbo build.

Signed-off-by: Icenowy Zheng <uwu@icenowy.me>